Make some parameters be per destination/client.
Currently, the following parameters are per destination/client:
- version
- community
- timeout
- retries

The remaining parameters are still per destination:
- max_packets_on_the_wire
- max_request_packet_size
- min_interval
- max_repetitions

Of those four, "max_request_packet_size" and "max_repetitions"
are used very "locally" and thus can also be made per
destination/client, but it is doubtful whether it
makes sense to do so.

The "max_packets_on_the_wire" and "min_interval" control
throttling and sharing of the queries towards the destination
irrespective of the number of clients, and thus HAVE
to stay per destination.

Thanks to Lars for discussion which cleared this up.