Malformed SNMP response bugfix:
There was a bug when processing a malformed SNMP response.
When the packet was "good enough" so that we could determined
the SID it corresponds to, but not good enough to finish SID
processing, the SID structure was unconditionally "forgotten",
so no retries or timeouts ever applied to it.  This lead to
the requesting client basically hanging waiting for a response
from SQE indefinitely.